refactor(ui): extract function getClosestYear into helpers
